+++ tests/types/public-surface.types.ts
@@ -123,6 +123,9 @@ injectedControlState.hasValidator(requiredReference, {});
 // @ts-expect-error Resolution must be a boolean.
 injectedControlState.hasValidator(requiredReference, { resolve: 'yes' });
 
+type _ShortCommittedOutput = Expect<Equal<typeof nameBinding.formNodeChange, OutputRef<string>>>;
+nameBinding.formNodeChange.subscribe(value => value.toUpperCase());
+
 type _CommittedOutput = Expect<Equal<typeof nameBinding.formNodeValueChange, OutputRef<string>>>;
 type _ControlOutput = Expect<Equal<typeof nameBinding.formNodeControlValueChange, OutputRef<string>>>;
 nameBinding.formNodeValueChange.subscribe(value => value.toUpperCase());